pyproject.toml
[project]
name = "mcpforge"
version = "0.2.0"
description = "Generate complete, runnable FastMCP 3.x MCP servers from plain-English descriptions"
authors = [
    { name = "saagpatel", email = "saagarpatel08@gmail.com" },
]
requires-python = ">=3.12"
license = { text = "MIT" }
readme = "README.md"
keywords = ["mcp", "fastmcp", "anthropic", "claude", "code-generation", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
mcpforge = "mcpforge.cli:cli"

[project.urls]
Homepage = "https://github.com/saagpatel/mcpforge"
Repository = "https://github.com/saagpatel/mcpforge"
"Bug Tracker" = "https://github.com/saagpatel/mcpforge/issues"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["mcpforge"]
